server/requirements.txt
fastapi
uvicorn
requests
python-dotenv
cerebras-cloud-sdk

tree-sitter
tree-sitter-python
tree-sitter-javascript
tree-sitter-typescript
tree-sitter-java
tree-sitter-go
tree-sitter-c
tree-sitter-cpp
tree-sitter-rust
tree-sitter-ruby
tree-sitter-php
tree-sitter-swift

pinecone

langchain
langchain-chroma
sentence-transformers
langchain_google_genai

# Socket.IO for async with FastAPI
python-socketio>=5.7.0
aiohttp

numpy>=1.21.0

requests